test: move wait_for_feature to pylib utils
The `wait_for_feature` can be useful, and will be used, in other test
suites than `topology_raft_disabled`, so it is moved to the common pylib
utils.
